refactor(R1): relocate Python-built UI into MorphoDepotLib
Move MorphoDepotAccessionForm, the six Form*Question widgets, MorphoDepotSearchForm,
and ScreenshotReviewDialog into MorphoDepotLib/{forms,accession_form,search_form,
screenshot_dialog}.py; main file imports them. MorphoDepot.py 7194 -> 6361 lines.
Net 11/11.
